LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/binary_leb128_64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "binary_leb128_64.wast:1"
PASS #3 Test that WebAssembly compilation fails
PASS #4 Reinitialize the default imports
FAIL #5 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 22: Memory64 is not enabled at {loc} expected true got false
FAIL #6 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
binary_leb128_64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/binary_leb128_64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/binary_leb128_64.wast.js:12:3 expected true got false
PASS #7 Test that WebAssembly compilation fails
